refactor: moving to download-file-cache.ts file
